QC now returns ArrayList<Integer>
Edited the QC method to return an array list of the indexes where banned words are found. This can easily be used to tell if any banned words are present by calling .size()>0 on the returned array list. The index returned is the start index of a given banned word.