Added auth, responses, and refactored
Added the authorization code to the endpoint methods, used the ResponseMessage class to send responses for endpoint methods, removed methods that were no longer in use, made other class methods private, and reformatted the code. The manual team endpoint is still available, but there are no current plans to implement this in the GUI with the time remaining in the project.